Update to use cffi 1.0 compilation mechanism
Create two ffi's for the module in ffi_build, one for the cairo library,
importing the ffi from xcffib (if it can be imported) and adding the
relevant cdefs, and another ffi for the gdk_pixbuf cdefs.

Now, to run cairocffi from source will require running `python
cairocffi/ffi_build.py` to compile the ffi objects, these are handled
automatically when installing.