Import built ffi module if globally installed
Import the built ffi modules from `cairocffi._ffi` rather that `._ffi`,
so that if the package has been installed, this will pick up the
compiled modules that are installed rather than insisting they be re-built
in the source directory. This has the unfortunate side-effect of possibly
hiding any problems if the cdefs are changed but the package is not
re-built in the source tree.