Keep cairo library loaded until all relevant objects are freed
The CFFI documentation says that the "library" object returned by
ffi.dlopen() is closed when it goes out of scope, and notes "Make sure
you keep the library object around as long as needed."

When a symbol is retrieved from a library, the returned object does not
keep a reference to the library it was loaded from.  It is possible for
the library to be closed while there are still some of these objects
around; using them subsequently can cause undefined behaviour.

When the Python interpreter is shutting down, the lack of reference from
the library symbol object to the library may cause it to destroy objects
in the wrong order; in particular, the cairo module may be wiped
(removing the only reference to the cairo library object) while cdata
objects returned by ffi.gc() with a destructor implemented in the cairo
library still exist.

This commit keeps a reference to the cairo library object alongside the
destructor, ensuring that the library will not be unloaded until all the
destructors have been called.  This should fix the segfault on exit
problem described in github issue #73.