Use Ubuntu Trusty image, test on Python 3.5
The version of Ubuntu that ships with Trusty (14.04) is _much_ better at
handling cairo/xcb stuff. It is still in beta, but it may be good to
switch to that to get all of the xcb tests up and running on Travis.
Also add Python 3.5 tests for Travis and tox.