Build generated files manually when building wheels
Source package must not include generated files, as they are different
depending on the target system, but they must be included in the wheel created
when the source package is installed.

We use PEP 517’s hooks to remove the generated files if they’re present, and to
build them during the wheel creation’s step of the installation.